Test concurrent registrations of users in a number of chats
(Specified by "Improvement 1" of exercise specification)

Use a CompletionService to which we submit tasks, each one simulating
one user. Each task registers a user in the chat manager and registers
the users in a number of chat rooms. Tasks are executed concurrently
in the thread pool.

If any exception is thrown in any of the tasks' thread, it will be raised
when calling Future.get, and the test will fail.