Verify that only one Chat instance is created for a given chat name
Even if a number of users are creating a chats with the same name
concurrently, only one instance with the chat name shall be created.

To verify this, we create a User mock and register it in the ChatManager.
At the end of the test, we verify that the number of calls to newChat
is equal to the number of chats that shall be instantiated.