Test case for basic chat operations
Being aware that, at some point, callbacks to User instances will
be asynchronous, i.e., will be run in a different thread than thread
running the test code.

We will have an array of TestUser with the implementation of the
callback modified so it counts down in a CountDownLatch. The
CountDownLatch is initialized with the number of mock users that
will be notified. After performing operation in the ChatManager,
the test thread will await for the count in the latch to be 0,
meaning that all users have been notified (each thread running a
callback will have decreased the count). Then the tests runs some
verifications on how callbacks have been invoked.

For call verifications, TestUser instances are wrapped with the
spy utility provided by Mockito, so we can verify that callbacks
have been performed with the right arguments