Execute all User callbacks asynchronously
Add to the UserCallbackHandler functions to execute asynchronously those
callbacks invoked by the ChatManager.

Let ChatManger have UserCallbackHandler instances rather than User
instances in its map.

The Chat insance gets the UserCallbackHandler instance for a certain
User from the ChatManger. This imposes the prerrequisite that a User
must be registered in the ChatManager before being added to a Chat,
otherwise a NullPointerException will be thrown.