Test registration of Users with same name in ChatManager
If a User with the same name already exists an exception must be thrown.
When several users try concurrently, only one shall succeed. For this to
work we have to use ConcurrentHashMap.compute method to ensure that
processing of the mapping logic is executed atomically.

Added tests to verify this behavior.