New feature: max wait time for chat creation
Added a BlockingQueue of `tokens`. The ChatManager initializes the queue
with a number of tokens equals to the max number of chats supported.

Before creating a chat, a token needs to be acquired. If the max number
of chats has been reached, the thread will be blocked at the BlockingQueue
of tokens for a new one to be available.

When a chat is removed, a new token is added to the queue, to flag that
capacity has increased. This will wake up one thread that might be waiting.
To ensure that concurrent removals of a chat do not over-increase the number
of tokens, we need to use chats.computeIfPresent, in order to remove the chat
and increase the capacity atomically.